feat: add get command to tptctl (#36)
* feat: add get command to tptctl


* feat: add get workload-instances command to tptctl


* feat: add get workloads command to tptctl


---------

+++ pkg/config/v0/workload.go
@@ -1,6 +1,7 @@
 package v0
 
 import (
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -92,16 +93,26 @@ func (w *WorkloadValues) Delete(apiClient *http.Client, apiEndpoint string) (*v0
 		return nil, nil, fmt.Errorf("failed to find workload instance with name %s: %w", workloadInstName, err)
 	}
 
-	// delete workload instance
-	deletedWorkloadInstance, err := client.DeleteWorkloadInstance(apiClient, apiEndpoint, *workloadInstance.ID)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to delete workload instance from threeport API: %w", err)
-	}
-
 	// get workload definition by name
 	workloadDefinition, err := client.GetWorkloadDefinitionByName(apiClient, apiEndpoint, w.Name)
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to find workload definition with name %s: %w", w.Name, err)
+	}
+
+	// ensure the workload definition has no more than one associated instance
+	workloadDefInsts, err := client.GetWorkloadInstancesByWorkloadDefinitionID(apiClient, apiEndpoint, *workloadDefinition.ID)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to get workload instances by workload definition with ID: %d: %w", workloadDefinition.ID, err)
+	}
+	if len(*workloadDefInsts) > 1 {
+		err = errors.New("deletion using the workload abstraction is only permitted when there is a one-to-one workload definition and workload instance relationship")
+		return nil, nil, fmt.Errorf("the workload definition has more than one workload instance associated: %w", err)
+	}
+
+	// delete workload instance
+	deletedWorkloadInstance, err := client.DeleteWorkloadInstance(apiClient, apiEndpoint, *workloadInstance.ID)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to delete workload instance from threeport API: %w", err)
 	}
 
 	// delete workload definition
